v1.45.4
I had originally tried to upgrade to 1.46 but this version does not
appear to work on heroku, so it failed in our staging environment and
was never deployed to production.

This commit is going to change the version to 1.45.4 which is the latest
version of the one we are on (in staging and production) and fixes the
security hole that we are trying to fix.