Use explicit URL path for metabase jar downloading
Up until this point, metabase had a convention to downloading the jar
that was predictable.

Given this is a high priority security patch, the url slug is different
requiring changes in how it is interpolated for downloading.

Going forward, we can continue to use this convention, but instead of
`1.45.5` it will be `enterprise/v1.45.5` which seems okay.